feat: Add @azure/storage-blob dependency for Azure Blob Storage integration
Add the "@azure/storage-blob" dependency with version "^12.24.0" to the package.json file. This dependency is required for integrating with Azure Blob Storage.

Also, add a new route "/edit-profile-image" in the userRoute.js file to handle the functionality of editing the user's profile image. The route is protected and can only be accessed by authenticated users. The editProfileImage function is responsible for uploading the image to Azure Blob Storage using the uploadImageStreamed function from the azureStorage.js file. The function also extracts metadata from the request headers using the extractMetadata function from the global.js file.

This commit addresses the following tasks:
- Add @azure/storage-blob dependency
- Implement /edit-profile-image route for uploading profile images